requirements-dev.txt
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.9

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# CDK testing
aws-cdk.assertions==2.120.0
